Fix zombie spawning back to on_player_died
An issue was hit pre-go live and so it zombie spawning was temporarily changed to be on pre_player_died as was a simple fix. But other mods could return player's health there and then they wouldn't die. So now we react to when the player dies and find their corpse again.