micro
Loading a save from an older zombie mod version that had testing of "true" caused an error for joining clients. The "true" value was a mistake in the 1.0.4 release.
Fix is to load the save with the new mod version, then save it to make a good save. Then load that save on the server and all will be fine.
The issue is due to the update order being Load(), then ConfigChanged() on the host. Meaning it uses the old value when deciding if to register the command, then updating the value. The client gets the current (updated) value on its connection and uses this when OnLoad().
Error:
4.637 Error ClientMultiplayerManager.cpp:1219: zombie_engineer was registered for the following commands when the map was saved but has not registered for any commands as a result of loading: zombie-engineer-call-on-startup
  94.637 Error ClientMultiplayerManager.cpp MultiplayerManager failed: "" + multiplayer.script-command-mismatch + "" + "zombie_engineer"